requirements.txt
buildozer==1.2.0
Cython==0.29.28
Kivy==2.1.0
kivymd==1.0.2
plyer==2.0.0
pyjnius==1.4.1
pytube==12.0.0

